Bugfix/issue 95 worker mem leak 2 (#106)
* Refactored to remove the ClusterMessages module altogether - it was adding too much complexity and memory overhead for only a slight improvement in readability over just using the Cluster module directly as necessary.

* Remove file ref count logic from CacheFS module. Memory usage was scaling with project size which eventually resulted in unacceptable memory usage. Also fixed a couple of warnings and errors in the test suite.
