+++ lib/cache/cache_fs.js
@@ -9,25 +9,8 @@ const moment = require('moment');
 const pick = require('lodash').pick;
 const crypto = require('crypto');
 const fileExtensions = require('lodash').invert(consts.FILE_TYPE);
-const cluster = require('cluster');
-const cm = require('../cluster_messages');
-
-const k_incrementGuidRef = "_incrementGuidRef";
-const k_releaseGuidRef = "_releaseGuidRef";
 
 class CacheFS extends CacheBase {
-    constructor() {
-        super();
-        this._guidRefs = {};
-
-        cm.listenFor(k_incrementGuidRef, key => {
-            this._incrementGuidRef(key);
-        });
-
-        cm.listenFor(k_releaseGuidRef, key => {
-            this._releaseGuidRef(key);
-        });
-    }
 
     get _optionsPath() {
         return super._optionsPath + ".cache_fs";
@@ -67,39 +50,6 @@ class CacheFS extends CacheBase {
 
     /**
      *
-     * @param {String} key
-     * @private
-     */
-    _incrementGuidRef(key) {
-        if(cluster.isWorker) {
-            return cm.send(k_incrementGuidRef, key);
-        }
-
-        if(this._guidRefs.hasOwnProperty(key)) {
-            this._guidRefs[key]++;
-        }
-        else {
-            this._guidRefs[key] = 1;
-        }
-    }
-
-    /**
-     *
-     * @param {String} key
-     * @private
-     */
-    _releaseGuidRef(key) {
-        if(cluster.isWorker) {
-            return cm.send(k_releaseGuidRef, key);
-        }
-
-        if(this._guidRefs.hasOwnProperty(key)) {
-            this._guidRefs[key]--;
-        }
-    }
-
-    /**
-     *
      * @param {String} type
      * @param {Buffer} guid
      * @param {Buffer} hash
@@ -123,10 +73,6 @@ class CacheFS extends CacheBase {
     async _writeFileToCache(type, guid, hash, sourcePath) {
         const filePath = this._calcFilepath(type, guid, hash);
 
-        if(this._guidRefs.hasOwnProperty(filePath) && this._guidRefs[filePath] > 0) {
-            throw new Error(`File is busy, cannot overwrite ${key}`);
-        }
-
         await fs.move(sourcePath, filePath, { overwrite: true });
         return filePath;
     }
@@ -143,13 +89,10 @@ class CacheFS extends CacheBase {
 
         return new Promise((resolve, reject) => {
             stream.on('open', () => {
-                this._incrementGuidRef(key);
                 resolve(stream);
             }).on('error', err => {
                 helpers.log(consts.LOG_ERR, err);
                 reject(err);
-            }).on('close', () => {
-                this._releaseGuidRef(key);
             });
         });
     }
