+++ lib/cache/reliability_manager.js
@@ -1,7 +1,6 @@
 const cluster = require('cluster');
 const path = require('path');
 const helpers = require('../helpers');
-const cm = require('../cluster_messages');
 const consts = require('../constants');
 const crypto = require('crypto');
 const _ = require('lodash');
@@ -15,6 +14,8 @@ const defaultOptions = {
 
 class ReliabilityManager {
     constructor(db, cachePath, options) {
+        this._id = ++ReliabilityManager._idCounter;
+        this._kMsgUpdateReliabilityFactor = `_updateReliabilityFactorForVersion.${this._id}`;
         this._db = db;
         this._cachePath = cachePath;
         this._options = options || {};
@@ -30,9 +31,13 @@ class ReliabilityManager {
             }
         }
 
-        cm.listenFor("_updateReliabilityFactorForVersion", (data) => {
-            return this._updateReliabilityFactorForVersion(data);
-        });
+        if(cluster.isMaster) {
+            cluster.on('message', (worker, msg) => {
+                if(msg._msg === this._kMsgUpdateReliabilityFactor) {
+                    return this._updateReliabilityFactorForVersion(msg);
+                }
+            });
+        }
     }
 
     /**
@@ -43,7 +48,8 @@ class ReliabilityManager {
      */
     async _updateReliabilityFactorForVersion(params) {
         if(cluster.isWorker) {
-            return cm.send('_updateReliabilityFactorForVersion', params);
+            params._msg = this._kMsgUpdateReliabilityFactor;
+            return process.send(params);
         }
 
         const entry = this.getEntry(params.guidStr, params.hashStr, true);
@@ -159,5 +165,7 @@ ReliabilityManager.reliabilityStates = {
     ReliableNew: 'ReliableNew',
     Reliable: 'Reliable'
 };
+
+ReliabilityManager._idCounter = 0;
 
 module.exports = ReliabilityManager;
