+++ lib/server/command_processor.js
@@ -113,7 +113,7 @@ class CommandProcessor extends Duplex {
         if(this[kReadStream] === null || this[kSource] === null) return;
 
         let chunk;
-        while(chunk = this[kReadStream].read()) {
+        while((chunk = this[kReadStream].read()) !== null) {
             this._sendFileQueueChunkReads++;
             this._sendFileQueueReadBytes += chunk.length;
             if(!this.push(chunk, 'ascii')) break;
