+++ test/test_init.js
@@ -1,4 +1,5 @@
 const helpers = require('../lib/helpers');
+require('cluster').setMaxListeners(25);
 
 process.on('unhandledRejection', (reason) => {
     console.error(reason);
