+++ test/test_utils.js
@@ -124,7 +124,7 @@ exports.readStream = function(stream, size) {
         const buffer = Buffer.alloc(size, 0, 'ascii');
         stream.on('readable', function() {
             let data;
-            while(data = this.read()) {
+            while((data = this.read()) !== null) {
                 if(pos + data.length > size) {
                     reject(new Error("Stream size exceeds buffer size allocation"));
                 }
